pyproject.toml
[project]
name = "SentientSearch"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "salzubi", email = "salaheddin@sentient.xyz"},
]
dependencies = [
    "aiohappyeyeballs==2.3.4",
    "aiohttp==3.10.0",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "attrs==23.2.0",
    "beautifulsoup4==4.12.3",
    "boto3==1.34.152",
    "botocore==1.34.152",
    "certifi==2024.7.4",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cohere==5.6.2",
    "colorama==0.4.6",
    "colorlog==6.8.2",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "fastapi==0.112.0",
    "fastavro==1.9.5",
    "filelock==3.15.4",
    "frozenlist==1.4.1",
    "fsspec==2024.6.1",
    "groq==0.9.0",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.24.5",
    "idna==3.7",
    "jmespath==1.0.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.2.12",
    "langchain-community==0.2.10",
    "langchain-core==0.2.27",
    "langchain-text-splitters==0.2.2",
    "langsmith==0.1.96",
    "marshmallow==3.21.3",
    "multidict==6.0.5",
    "mypy-extensions==1.0.0",
    "numpy==1.26.4",
    "openai==1.38.0",
    "orjson==3.10.6",
    "packaging==24.1",
    "parameterized==0.9.0",
    "pydantic==2.8.2",
    "pydantic-core==2.20.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.1",
    "regex==2023.12.25",
    "requests==2.32.3",
    "requests-mock==1.12.1",
    "s3transfer==0.10.2",
    "semantic-chunkers==0.0.9",
    "semantic-router==0.0.55",
    "six==1.16.0",
    "sniffio==1.3.1",
    "soupsieve==2.5",
    "SQLAlchemy==2.0.31",
    "starlette==0.37.2",
    "tenacity==8.5.0",
    "tiktoken==0.7.0",
    "tokenizers==0.19.1",
    "tqdm==4.66.4",
    "types-requests==2.32.0.20240712",
    "typing-inspect==0.9.0",
    "typing-extensions==4.12.2",
    "urllib3==2.2.2",
    "uvicorn==0.30.5",
    "uvloop==0.19.0",
    "watchfiles==0.22.0",
    "websockets==12.0",
    "yarl==1.9.4",
    "ipykernel>=6.29.5",
    "ipdb>=0.13.13",
    "pandas>=2.2.3",
    "jinja2>=3.1.4",
    "docling>=2.7.1",
    "fireworks-ai>=0.15.8",
    "timeout-decorator>=0.5.0",
    "markdownify>=0.14.1",
    "readability-lxml>=0.8.1",
    "lxml[html_clean]>=5.3.0",
    "joblib>=1.4.2",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
